Project/Backend/server/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.6.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
